refactor: move GitHub related plugin together
