--- app/plugin/github-webhook-manager/GitHubWebhook.ts
@@ -1,89 +1,0 @@
-// Copyright 2019 Xlab
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-import { Application, Context } from 'egg';
-import Webhooks from '@octokit/webhooks';
-import EventSource from 'eventsource';
-import { AppPluginBase } from '../../basic/AppPluginBase';
-import { InstallationInitEvent } from '../installation-manager/types';
-import { GitHubClientConfig } from '../github-client-manager/GitHubInstallationConfig';
-import { join } from 'path';
-import { GitHubWebhooksManagerReadyEvent } from '../event-manager/events';
-
-export class GitHubWebhook extends AppPluginBase<null> {
-
-  private webhooks: Map<number, Webhooks>;
-
-  constructor(config: null, app: Application) {
-    super(config, app);
-    this.webhooks = new Map<number, Webhooks>();
-  }
-
-  public async onReady(): Promise<void> {
-    this.app.event.subscribeAll(InstallationInitEvent, async e => {
-      // init webhooks for each app
-      const config: GitHubClientConfig = e.config;
-      const webhooks = new Webhooks({
-        secret: config.webhook.secret,
-      });
-      this.webhooks.set(e.installationId, webhooks);
-
-      // setup proxy using smee.io
-      if (config.webhook.proxyUrl) {
-        const source = new EventSource(config.webhook.proxyUrl);
-        source.onmessage = (event: any) => {
-          const webhookEvent = JSON.parse(event.data);
-          webhooks.verifyAndReceive({
-            id: webhookEvent['x-request-id'],
-            name: webhookEvent['x-github-event'],
-            signature: webhookEvent['x-hub-signature'],
-            payload: webhookEvent.body,
-          }).catch(this.logger.error);
-        };
-      }
-
-      // setup router
-      const path = join(config.webhook.path, e.installationId.toString());
-      this.logger.info(`Load webhooks for ${config.name} on /${this.name}${path}`);
-      this.get(path, async (ctx: Context, next: any) => {
-        // pass to webhooks
-        webhooks.verifyAndReceive({
-          id: ctx.headers['x-request-id'],
-          name: ctx.headers['x-github-event'],
-          signature: ctx.headers['x-hub-signature'],
-          payload: ctx.body,
-        }).catch(this.logger.error);
-        ctx.body = 'ok';
-        await next();
-      });
-
-      this.logger.info('Start send wehbook ready to ', e.installationId);
-      this.app.event.publish('worker', GitHubWebhooksManagerReadyEvent, {
-        installationId: e.installationId,
-      });
-    });
-  }
-
-  public async onStart(): Promise<void> { }
-
-  public async onClose(): Promise<void> { }
-
-  public register(installationId: number, func: (webhooks: Webhooks) => any) {
-    const webhooks = this.webhooks.get(installationId);
-    if (webhooks) {
-      func(webhooks);
-    }
-  }
-
-}
