--- app/plugin/github-webhook-manager/package.json
@@ -1,7 +1,0 @@
-{
-  "name": "mentor-bot-github-webhook",
-  "description": "GitHub webhook for mentor bot",
-  "eggPlugin": {
-    "name": "githubWebhook"
-  }
-}
